+++ jest.config.js
@@ -4,7 +4,7 @@ const { compilerOptions } = require('./tsconfig.json');
 module.exports = {
 	clearMocks: true,
 	collectCoverage: true,
-	collectCoverageFrom: ['<rootDir>/src/modules/**/services/*.ts'],
+	collectCoverageFrom: ['<rootDir>/src/**/**/services/*.ts'],
 	coverageDirectory: 'coverage',
 	coverageReporters: ['text-summary', 'lcov'],
 	globalSetup: './jest.setup.js',
