+++ src/modules/categories/application/services/create-category.service.ts
@@ -1,8 +1,11 @@
 import { CreateCategoryDto } from '@modules/categories/domain/dto/create-category.dto';
 import { ICategoriesRepository } from '@modules/categories/domain/interfaces/category-repository.interface';
+
+import { CreateUserDto } from '@modules/users/domain/dto/create-user.dto';
+import { ConflictError } from '@shared/errors/conflict.error';
 import { Category } from '@modules/categories/infra/persistence/typeorm/entities/category.entity';
 import { inject, injectable } from 'tsyringe';
-import { ConflictError } from '@shared/errors/conflict.error';
+
 import slugify from 'slugify';
 
 @injectable()
